Moved stores and styles
